Fix license handling for non-English languages
dmgbuild was using the language codes corresponding to langXXX from
Script.h... it should actually be using the region codes corresponding
to verXXX from Script.h, which it does now.

This fixes a problem which caused nearly every language but English
(which happens to use '0' for both langEnglish and verUS) to show
garbled or broken text or display in the wrong language.

This also changes dmgbuild to use ISO language and country identifiers
to identify locales, rather than arbitrary values based on the names of
constants in Script.h (which were incorrect anyways).

dmgbuild also no longer lists encodings that apparently don't exist, and
instead displays an informative error for unsupported encodings.

TMPL resources are no longer written either, and they do not appear to
be necessary. They are apparently only used by graphical resource
editors which have long since fallen out of use (and so there is no
point adding it when all the data is auto-generated).

Setting the default language now also works, as the key being used was
'default-lang' contrary to 'default-language' as specified in the
documentation.